💣 Pull apart giant frontmatter file
