+++ packages/myst-frontmatter/src/contributors/index.ts
@@ -1,0 +1,2 @@
+export * from './types.js';
+export * from './validators.js';
