+++ packages/myst-frontmatter/src/exports/index.ts
@@ -1,0 +1,2 @@
+export * from './types.js';
+export * from './validators.js';
