--- packages/myst-frontmatter/src/frontmatter/index.ts
@@ -1,2 +1,0 @@
-export * from './types.js';
-export * from './validators.js';
