+++ packages/myst-frontmatter/src/index.ts
@@ -1,2 +1,13 @@
-export * from './frontmatter/index.js';
+export * from './affiliations/index.js';
+export * from './biblio/index.js';
+export * from './contributors/index.js';
+export * from './exports/index.js';
+export * from './funding/index.js';
 export * from './licenses/index.js';
+export * from './numbering/index.js';
+export * from './page/index.js';
+export * from './project/index.js';
+export * from './site/index.js';
+export * from './thebe/index.js';
+export * from './utils/index.js';
+export * from './venues/index.js';
