+++ packages/myst-frontmatter/src/numbering/types.ts
@@ -1,0 +1,13 @@
+export type Numbering = {
+  enumerator?: string;
+  figure?: boolean;
+  equation?: boolean;
+  table?: boolean;
+  code?: boolean;
+  heading_1?: boolean;
+  heading_2?: boolean;
+  heading_3?: boolean;
+  heading_4?: boolean;
+  heading_5?: boolean;
+  heading_6?: boolean;
+};
