+++ packages/myst-frontmatter/src/project/validators.ts
@@ -1,0 +1,230 @@
+import type { ValidationOptions } from 'simple-validators';
+import {
+  defined,
+  filterKeys,
+  incrementOptions,
+  validateBoolean,
+  validateDate,
+  validateList,
+  validateObject,
+  validateObjectKeys,
+  validateString,
+  validateUrl,
+} from 'simple-validators';
+import { validateBiblio } from '../biblio/validators.js';
+import { validateExportsList } from '../exports/validators.js';
+import { validateLicenses } from '../licenses/validators.js';
+import { validateNumbering } from '../numbering/validators.js';
+import { KNOWN_PAGE_ALIASES } from '../page/validators.js';
+import { SITE_FRONTMATTER_KEYS, validateSiteFrontmatterKeys } from '../site/validators.js';
+import { validateThebe } from '../thebe/validators.js';
+import { validateBibliography, validateBooleanOrObject, validateDoi } from '../utils/validators.js';
+import type { ProjectFrontmatter } from './types.js';
+
+export const PROJECT_AND_PAGE_FRONTMATTER_KEYS = [
+  'date',
+  'name',
+  'doi',
+  'arxiv',
+  'open_access',
+  'license',
+  'binder',
+  'source',
+  'subject',
+  'biblio',
+  'oxa',
+  'numbering',
+  'bibliography',
+  'math',
+  'abbreviations',
+  'exports',
+  // Do not add any project specific keys here!
+  ...SITE_FRONTMATTER_KEYS,
+];
+
+// These keys only exist on the project.
+export const PROJECT_FRONTMATTER_KEYS = [
+  ...PROJECT_AND_PAGE_FRONTMATTER_KEYS,
+  'references',
+  'requirements',
+  'resources',
+  'thebe',
+];
+
+const KNOWN_PROJECT_ALIASES = {
+  ...KNOWN_PAGE_ALIASES,
+  // This must also be updated in myst-config
+  jupyter: 'thebe',
+};
+
+export function validateSharedProjectFrontmatterKeys(
+  value: Record<string, any>,
+  opts: ValidationOptions,
+) {
+  const output: Omit<ProjectFrontmatter, 'thebe' | 'resources' | 'requirements' | 'references'> =
+    validateSiteFrontmatterKeys(value, opts);
+  if (defined(value.date)) {
+    output.date = validateDate(value.date, incrementOptions('date', opts));
+  }
+  if (defined(value.name)) {
+    output.name = validateString(value.name, incrementOptions('name', opts));
+  }
+  if (defined(value.doi)) {
+    output.doi = validateDoi(value.doi, incrementOptions('doi', opts));
+  }
+  if (defined(value.arxiv)) {
+    output.arxiv = validateUrl(value.arxiv, {
+      ...incrementOptions('arxiv', opts),
+      includes: 'arxiv.org',
+    });
+  }
+  if (defined(value.open_access)) {
+    output.open_access = validateBoolean(value.open_access, incrementOptions('open_access', opts));
+  }
+  if (defined(value.license)) {
+    output.license = validateLicenses(value.license, incrementOptions('license', opts));
+  }
+  if (defined(value.binder)) {
+    output.binder = validateUrl(value.binder, incrementOptions('binder', opts));
+  }
+  if (defined(value.source)) {
+    output.source = validateUrl(value.source, incrementOptions('source', opts));
+  }
+  if (defined(value.subject)) {
+    output.subject = validateString(value.subject, {
+      ...incrementOptions('subject', opts),
+      maxLength: 40,
+    });
+  }
+  if (defined(value.bibliography)) {
+    output.bibliography = validateBibliography(
+      value.bibliography,
+      incrementOptions('bibliography', opts),
+    );
+  }
+  if (defined(value.biblio)) {
+    output.biblio = validateBiblio(value.biblio, incrementOptions('biblio', opts));
+  }
+  if (defined(value.oxa)) {
+    // TODO: better oxa validation
+    output.oxa = validateString(value.oxa, incrementOptions('oxa', opts));
+  }
+  if (defined(value.numbering)) {
+    output.numbering = validateBooleanOrObject(
+      value.numbering,
+      incrementOptions('numbering', opts),
+      validateNumbering,
+    );
+  }
+  if (defined(value.math)) {
+    const mathOpts = incrementOptions('math', opts);
+    const math = validateObject(value.math, mathOpts);
+    if (math) {
+      const stringKeys = Object.keys(math).filter((key) => {
+        // Filter on non-string values
+        return validateString(math[key], incrementOptions(key, mathOpts));
+      });
+      output.math = filterKeys(math, stringKeys);
+    }
+  }
+  if (defined(value.abbreviations)) {
+    const abbreviationsOpts = incrementOptions('abbreviations', opts);
+    const abbreviations = validateObject(value.abbreviations, abbreviationsOpts);
+    if (abbreviations) {
+      const stringKeys = Object.keys(abbreviations).filter((key) => {
+        // Filter on non-string values
+        return validateString(abbreviations[key], incrementOptions(key, abbreviationsOpts));
+      });
+      output.abbreviations = filterKeys(abbreviations, stringKeys);
+    }
+  }
+  if (defined(value.exports)) {
+    const exports = validateExportsList(value.exports, opts);
+    if (exports) output.exports = exports;
+  }
+  if (value.thumbnail === null) {
+    // It is possible for the thumbnail to explicitly be null.
+    // This means not to look to the images in a page.
+    output.thumbnail = null;
+  } else if (defined(value.thumbnail)) {
+    output.thumbnail = validateString(value.thumbnail, incrementOptions('thumbnail', opts));
+  }
+  if (defined(value.thumbnailOptimized)) {
+    // No validation, this is expected to be set programmatically
+    output.thumbnailOptimized = value.thumbnailOptimized;
+  }
+  if (value.banner === null) {
+    // It is possible for the banner to explicitly be null.
+    // This means not to look to the images in a page.
+    output.banner = null;
+  } else if (defined(value.banner)) {
+    output.banner = validateString(value.banner, incrementOptions('banner', opts));
+  }
+  if (defined(value.bannerOptimized)) {
+    // No validation, this is expected to be set programmatically
+    output.bannerOptimized = value.bannerOptimized;
+  }
+  return output;
+}
+
+export function validateProjectFrontmatterKeys(
+  value: Record<string, any>,
+  opts: ValidationOptions,
+) {
+  const output: ProjectFrontmatter = validateSharedProjectFrontmatterKeys(value, opts);
+  // This is only for the project, and is not defined on pages
+  if (defined(value.references)) {
+    const referencesOpts = incrementOptions('references', opts);
+    const references = validateObject(value.references, referencesOpts);
+    if (references) {
+      output.references = Object.fromEntries(
+        Object.keys(references)
+          .map((key) => {
+            const url = validateUrl(references[key], incrementOptions(key, referencesOpts));
+            if (!url) return undefined;
+            return [key, { url }];
+          })
+          .filter((exists) => !!exists) as [string, { url: string }][],
+      );
+    }
+  }
+
+  if (defined(value.thebe)) {
+    const result = validateThebe(value.thebe, incrementOptions('thebe', opts));
+    if (result && Object.keys(result).length > 0) output.thebe = result;
+    else delete output.thebe;
+  }
+
+  if (defined(value.requirements)) {
+    output.requirements = validateList(
+      value.requirements,
+      incrementOptions('requirements', opts),
+      (req, index) => {
+        return validateString(req, incrementOptions(`requirements.${index}`, opts));
+      },
+    );
+  }
+  if (defined(value.resources)) {
+    output.resources = validateList(
+      value.resources,
+      incrementOptions('resources', opts),
+      (res, index) => {
+        return validateString(res, incrementOptions(`resources.${index}`, opts));
+      },
+    );
+  }
+  return output;
+}
+
+/**
+ * Validate ProjectFrontmatter object against the schema
+ */
+export function validateProjectFrontmatter(input: any, opts: ValidationOptions) {
+  const value =
+    validateObjectKeys(
+      input,
+      { optional: PROJECT_FRONTMATTER_KEYS, alias: KNOWN_PROJECT_ALIASES },
+      opts,
+    ) || {};
+  return validateProjectFrontmatterKeys(value, opts);
+}
