+++ packages/myst-frontmatter/src/utils/index.ts
@@ -1,1 +1,6 @@
+export * from './fillPageFrontmatter.js';
+export * from './normalizeString.js';
 export * from './parseName.js';
+export * from './referenceStash.js';
+export * from './unnestKernelSpec.js';
+export * from './validators.js';
