+++ packages/myst-frontmatter/src/utils/normalizeString.ts
@@ -1,0 +1,7 @@
+export function normalizeJsonToString(value: Record<string, any>) {
+  return JSON.stringify(
+    Object.entries(value)
+      .filter(([, val]) => val !== undefined)
+      .sort(),
+  );
+}
