+++ packages/myst-frontmatter/src/venues/types.ts
@@ -1,0 +1,4 @@
+export type Venue = {
+  title?: string;
+  url?: string;
+};
